pyproject.toml
[tool.poetry]
name = "recruitment-fastapi"
version = "0.1.0"
description = "Recruitement project"
authors = ["dhara-impero"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.5.0"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
email-validator = "^2.1.0.post1"
python-dotenv = "^1.0.0"
fastapi-sso = "^0.7.2"
pydantic-settings = "^2.0.3"
uvloop = "^0.20.0"
httptools = "^0.6.1"
pandas = "^2.2.2"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
